% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Basics.R
\name{binary}
\alias{binary}
\title{Binary}
\usage{
binary(i, bit)
}
\arguments{
\item{i}{numeric; numeric to be transformed in a binaric array}

\item{bit}{numeric; number of digits for the binaric ountput}
}
\description{
Creates the Binary output of a number
}

