% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Spatial.R
\name{centroidMean}
\alias{centroidMean}
\title{Mean of Points within Shapefile}
\usage{
centroidMean(spat, pnt, count = T)
}
\arguments{
\item{spat}{SpatialPolygonDataFrame; A file containoing one or more Spatial
Polygons}

\item{pnt}{SpatialPointDataFrame; A file containing the Point cloud.}

\item{count}{boolean; If T, count the number of points within one Polygon. OPTIONAL}
}
\description{
The centroid Mean function returns a SpatialPointDataframe
containing the Centroid mean of Points within ESRI Shapefile(s).
}
