% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Initialization.R
\name{dircheckup}
\alias{dircheckup}
\title{dircheckup}
\usage{
dircheckup(mydir)
}
\arguments{
\item{mypkg}{character; Location of a Directory}
}
\description{
Automatically checks whether a Directory exists. If none is
available it will automatically create one.
By specifying the Subdirectory also non-existing main directories will
be automatically built
}

