% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Basics.R
\name{multigrepl}
\alias{multigrepl}
\title{Multiple grepl operations}
\usage{
multigrepl(mypatterns, x)
}
\arguments{
\item{mypatterns}{string(s); list of multiple input strings}

\item{x}{string(s); list of multiple target patterns}
}
\description{
Allows to add multiple patterns to a grepl operation.
These multiple patterns are then tested against other strings.
This function is based on the base grepl operator
}
