% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Basics.R
\name{startfun}
\alias{startfun}
\title{StartFun}
\usage{
startfun(myinp)
}
\arguments{
\item{myinp}{empty array or matrix; Thisobject must be empty or just created
filled with NA.}
}
\description{
Checks the first NA value within a matrix or array and returns
the location. Useful function for iteratively filling an object without
the use of another iteration.This function autmatically states the first NA value
}

