% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Spatial.R
\name{extract2}
\alias{extract2}
\title{Extract 2}
\usage{
extract2(rr, pp, points = 1000, samp.type = "regular", weight = F,
  sd = F, narm = F, seed = 1)
}
\arguments{
\item{rr}{Raster object; raster image}

\item{pp}{Spatial object; polygon file}

\item{points}{integer; Number of SpatialPoints created in the Polygon}

\item{weight}{boolean; Weighting image considering the underlaying rasters a
differentweight in the extraction computation. This makes pixel partially
located within a polygon less weighted. Input is T/F}

\item{sd}{boolean; computes the standard deviation of each extracted raster}

\item{narm}{boolean; removes na values}

\item{seed}{integer; applies the set.seed() function enabling the control over}

\item{sample.type}{character; Sampling technique for the Spatial Points
generated as described in the spsample() function of the "sp" package}
}
\description{
With this functions raster values are extracted by various
overlaying spatial points created within a polygon shapefile.
}

