% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Spatial.R
\name{extract2}
\alias{extract2}
\title{Vectorized Extraction of rasterValues}
\usage{
extract2(rr, pp, points = NA, samp.type = "regular", weight = F,
  narm = F, seed = 1, returnVals = F)
}
\arguments{
\item{rr}{Raster object; raster image}

\item{pp}{Spatial object; polygon file}

\item{points}{integer; Number of SpatialPoints created in the Polygon}

\item{samp.type}{character; Sampling technique for the Spatial Points
generated as described in the spsample() function of the "sp" package}

\item{weight}{boolean; Weighting image considering the underlaying rasters a
differentweight in the extraction computation. This makes pixel partially
located within a polygon less weighted. Input is T/F}

\item{narm}{boolean; removes na values}

\item{seed}{integer; applies the set.seed() function enabling the control over}

\item{returnVals}{boolean; return all extracted values}
}
\description{
With this functions raster values are extracted by various
overlaying spatial points created within a polygon shapefile.
}
