% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Initialization.R
\name{loadandinstall}
\alias{loadandinstall}
\title{Install and Initialize Packages}
\usage{
loadandinstall(mypkg)
}
\arguments{
\item{mypkg}{character; CRAN Package}
}
\description{
Loads and installs a CRAN hosted Package automatically in R.
No further need to apply both install.package() and library()
}
